GregMiscellaneous: zzogl-pg: Use some of the routines in the Utilities library instead of the copies in ZZOgl. (Which were basically older versions of the same code.)
git-svn-id: http://pcsx2.googlecode.com/svn/branches/GregMiscellaneous@4006 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/zzogl-pg/opengl/GSmain.cpp
@@ -536,7 +536,7 @@ void CALLBACK GSvsync(int interlace)
 	if (--nToNextUpdate <= 0)
 	{
 		u32 d = timeGetTime();
-		fFPS = UPDATE_FRAMES * 1000.0f / (float)max(d - dwTime, 1);
+		fFPS = UPDATE_FRAMES * 1000.0f / (float)max(d - dwTime, (u32)1);
 		dwTime = d;
 		g_nFrame += UPDATE_FRAMES;
 		SetGSTitle();
